¡Markov Model within Hash Map!
We have a markov model that is implemented within a hash map!

It can be trained from hard-coded input, but there are still heap
corruption issues as well as logistical bugs that still need to be
ironed out.

All of this work done on-stream at https://twitch.tv/lens_r